+++ tests/harness/__init__.py
@@ -1,0 +1,26 @@
+"""AirStack test harness — helpers split out of conftest.py by concern.
+
+conftest.py holds the pytest hooks and fixtures and imports what it needs from here.
+Public helpers are re-exported so tests and conftest can ``from harness import <name>``.
+"""
+from harness.discovery import (
+    AIRSTACK_ROOT,
+    COLCON_UNIT_TEST_PACKAGES_YAML,
+    colcon_test_robot_command,
+    load_colcon_unit_test_config,
+    repo_path,
+    unit_test_dirs,
+    unit_test_files,
+)
+from harness.session import logger
+
+__all__ = [
+    "AIRSTACK_ROOT",
+    "COLCON_UNIT_TEST_PACKAGES_YAML",
+    "colcon_test_robot_command",
+    "load_colcon_unit_test_config",
+    "repo_path",
+    "unit_test_dirs",
+    "unit_test_files",
+    "logger",
+]
